+++ monkey/monkey_island/cc/services/reporting/report_exporter_manager.py
@@ -1,15 +1,8 @@
 import logging
 
+from common.utils.code_utils import Singleton
+
 logger = logging.getLogger(__name__)
-
-
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
 
 
 class ReportExporterManager(object, metaclass=Singleton):
